# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ki-db Knowledge Index API
  version: 1.0.0
  description: |
    Postgres-centric Knowledge Index DB with hybrid BM25+vector search,
    knowledge graph, and event-driven indexing pipeline.
servers:
  - url: http://localhost:8080
    description: Ingest API
  - url: http://localhost:8081
    description: Query API
components:
  securitySchemes:
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
  schemas:
...
    SearchResult:
      type: object
      properties:
        chunk_id:
          type: string
          format: uuid
        doc_id:
          type: string
          format: uuid
        title:
          type: string
        heading_path:
          type: string
        chunk_text:
          type: string
        bm25_rank:
          type: integer
        ann_rank:
          type: integer
        rrf_score:
          type: number
        rerank_score:
          type: number
    ContextPack:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
    Citation:
      type: object
      properties:
        doc_id:
          type: string
          format: uuid
        version_id:
          type: string
          format: uuid
        title:
          type: string
        heading_path:
          type: string
        chunk_id:
          type: string
          format: uuid
        relevance:
          type: string
          enum: [keyword, semantic, both]
...
security:
  - TenantHeader: []
paths:
...
  /api/v1/search:
    get:
      tags: [Search]
      summary: Hybrid search (BM25 + vector ANN + optional rerank)
      operationId: search
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: rerank
          in: query
          schema:
            type: boolean
            default: false
        - name: doc_type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Search results with context pack
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextPack'
  /api/v1/feedback:
    post:
      tags: [Search]
      summary: Submit search feedback
      operationId: submitFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query, chunk_id, helpful]
              properties:
                query:
                  type: string
                chunk_id:
                  type: string
                  format: uuid
                helpful:
                  type: boolean
                note:
                  type: string
      responses:
        '201':
          description: Feedback recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  feedback_id:
                    type: string
                    format: uuid
...
  /api/v1/ops/jobs:
    get:
      tags: [Ops]
      summary: List recent indexing jobs
      operationId: listJobs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed]
      responses:
        '200':
          description: List of indexing jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      type: object
                      properties:
                        event_id:
                          type: string
                          format: uuid
                        event_type:
                          type: string
                        status:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                        published_at:
                          type: string
                          format: date-time

  # --- Health ---
...